+++ packages/class-core/CHANGELOG.md
@@ -10,6 +10,7 @@
   * annotations: `moveAndReplaceAtIndex[Directly]`
   * containments: `replace{With|AtIndex}`
   * references: `removeAtIndexDirectly`, `moveAndReplaceDirectly`
+* Add a `deltaHandlerForwardingTo` function that returns a delta handler function that forwards received deltas to the given delta handlers.
 
 
 ## 0.6.13